LPD-40735 - Refactor. Update interface reference
